added log models typeguards
