+++ src/logging/log-models.ts
@@ -11,12 +11,6 @@ export type LogLevelName = keyof typeof LogLevel;
 
 
 
-/**
- * Custom type guard to assert whether an unknown object is a LogLevelName.
- */
-export function isLogLevelName(maybeLogLevelName: unknown): maybeLogLevelName is LogLevelName {
-    return typeof maybeLogLevelName === 'string' && Object.keys(LogLevel).includes(maybeLogLevelName);
-}
 
 
 export const GravityLogLevelNames: Record<LogLevelName, string> = {
@@ -27,6 +21,7 @@ export const GravityLogLevelNames: Record<LogLevelName, string> = {
     ERROR: "Error",
     FATAL: "Fatal"
 } as const;
+
 export const RhinoLogLevelNames: Record<LogLevelName, string> = {
     DEBUG: "DBG",
     TRACE: "TRC",
@@ -45,20 +40,12 @@ export interface LogMessage {
     formattedMessage: string;
 }
 
-
-export function isLogMessage(object: any): object is LogMessage{
-    return Object.prototype.hasOwnProperty.call(object, "source")
-    && Object.prototype.hasOwnProperty.call(object, "level") 
-    && Object.prototype.hasOwnProperty.call(object, "message")
-    && Object.prototype.hasOwnProperty.call(object, "formattedMessage");
-}
-
 export interface GravityLogMessage extends LogMessage{
 }
 
 export interface RhinoLogMessage extends LogMessage{
     /**
-     * If loggerName exists, is of {application}.{loggerName}, otherwise equals to {application}.
+     * If loggerName exists, is {application}.{loggerName}, otherwise equals to {application}.
      */
     source: string;
     /**
